% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\docType{data}
\name{apa_partial_obj}
\alias{apa_partial_obj}
\title{American Psychological Association (APA) election data (partial rankings included)}
\format{a RankData object}
\usage{
apa_partial_obj
}
\description{
A dataset containing 5738 complete votes and 9711 partial votes in APA election. There are 5 candidates in total.
}
\references{
Marden, J. I. (1995). Analyzing and Modeling Rank Data (94-96). Chapman Hall, New York.
}
\seealso{
\code{\link{apa_obj}}
}
\keyword{datasets}

