% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{DistanceBlock}
\alias{DistanceBlock}
\title{Calculate Kendall distance
#' Calculate Kendall distance matrix between one ranking and a matrix of rankings}
\usage{
DistanceBlock(mat, r)
}
\arguments{
\item{mat}{a matrix of rankings}

\item{r}{a single ranking}
}
\value{
a vector of Kendall distance
}
\description{
Calculate Kendall distance
#' Calculate Kendall distance matrix between one ranking and a matrix of rankings
}

