% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{DistanceMatrix}
\alias{DistanceMatrix}
\title{Calculate Kendall distance matrix between rankings}
\usage{
DistanceMatrix(ranking)
}
\arguments{
\item{ranking}{a matrix of rankings}
}
\value{
Kendall distance matrix between rankings
}
\description{
Calculate Kendall distance matrix between rankings
}

