% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{DistancePair}
\alias{DistancePair}
\title{Calculate Kendall distance between a pair of rankings}
\usage{
DistancePair(r1, r2)
}
\arguments{
\item{r1}{a single ranking}

\item{r2}{a single ranking}
}
\value{
Kendall distance value
}
\description{
Calculate Kendall distance between a pair of rankings
}

