% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{GenerateExample}
\alias{GenerateExample}
\title{Generate simple examples}
\usage{
GenerateExample(ranking = TRUE)
}
\arguments{
\item{ranking}{TRUE if "ranking" representation is used in the output data; otherwise "ordering" representation is used.}
}
\description{
This function generates simple examples for illustrative proposes.
The sample contains the rankings of five objects and the underlying model is a Mallow's phi
model with dispersion parameter set to 0.2
and modal ranking set to (1,2,3,4,5)
}

