% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{GenerateExampleTopQ}
\alias{GenerateExampleTopQ}
\title{Generate simple examples of top-q}
\usage{
GenerateExampleTopQ()
}
\description{
This function generates simple examples for illustrative proposes.
The sample contains the top-3 rankings of five objects and the underlying model is a weighted Kendall distance model
model with weights set to (0.7,0.5,0.3,0)
and modal ranking set to (1,2,3,4,5)
}

