% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{HashtoRank}
\alias{HashtoRank}
\title{Obtain Ranking from Hash Value}
\usage{
HashtoRank(h)
}
\arguments{
\item{h}{A vector of hash values.}
}
\value{
a matrix of rankings if input has more than one element or a single ranking (numeric vector) if input has only one element
}
\description{
\code{HashToRank} returns rankings from given hash values. Maximum 52 objects are supported.
}
\seealso{
\code{\link{RanktoHash}} for a reverse operation.
}

