% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{ModelSummary}
\alias{ModelSummary}
\title{Print a brief summary of the fitted model}
\usage{
ModelSummary(model)
}
\arguments{
\item{model}{a ranking model returned by a call to RankDistanceModel function.}
}
\description{
Print a brief summary of the fitted model. This includes information about goodness
of fit as well as parameter estimation.
}

