% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\docType{class}
\name{RankControl-class}
\alias{RankControl}
\alias{RankControl-class}
\title{RankControl Class}
\description{
A S4 class to store control parameters for model fitting
}
\details{
The control parameters that start with prefix \code{EM_} are intended for the EM iteration. The ones with prefix \code{SeachPi0} control the behaviour of searching model ranking.
The function object specified in the \code{SearchPi0_control} takes a list as argument. The components in the list include the following. \code{obs}: the number of observations.
\code{w.est}: the estimated weights. \code{log_likelihood}: the estimated log_likelihood. With this information, most of the popular information criterions can be supported and customized criterions can also be defined.
A larger returned value indicates a better fit.
}
\section{Slots}{

\describe{
\item{\code{EM_limit}}{maximum number of EM iteration}

\item{\code{EM_epsilon}}{convergence error for weights and cluster probabilities in EM iteration}

\item{\code{SearchPi0_limit}}{maximum number of iterations in the local seach of pi0.}

\item{\code{SearchPi0_FUN}}{a function object that gives a goodness of fit criterion. The default is log likelihood.}

\item{\code{SearchPi0_fast_traversal}}{a logical value. If TRUE (by default), immediately traverse to the neighbour if it is better than the current pi0. Otherwise, check all neighbours and traverse to the best one.}

\item{\code{SearchPi0_show_message}}{a logical value. If TRUE, the location of the current pi0 is shown.}

\item{\code{SearchPi0_neighbour}}{a character string specifying which type of neighbour to use in the local search. Supported values are: "Cayley" to use neighbours in terms of Cayley distance or "Kendall" to use neighbours in terms of Kendall distance.
Note that Kendall neighbours are a subset of Cayley neighbours}

\item{\code{optimx_control}}{a list to be passed to \code{\link[optimx]{optimx}}. The list must not contain a component \code{maximize=TRUE} since internally the negation of the likelihood function is minimized.}
}}
\examples{
# enabling messages and warnings
testctrl = new("RankControl",SearchPi0_show_message=TRUE, optimx_control=list(dowarn=TRUE))
}
\seealso{
\code{\link{RankData}}, \code{\link{RankInit}}
}

