% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{RankControlCayley-class}
\alias{RankControlCayley}
\alias{RankControlCayley-class}
\title{RankControlCayley Class}
\description{
A S4 class for the Cayley distance model fitting.
It is derived from class \code{\link{RankControl-class}}.
}

