% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{RankControlFootrule-class}
\alias{RankControlFootrule}
\alias{RankControlFootrule-class}
\title{RankControlFootrule Class}
\description{
A S4 class for the Footrule distance model fitting.
It is derived from class \code{\link{RankControl-class}}.
}
\seealso{
\code{\link{RankData}}, \code{\link{RankInit}}, \code{\link{RankControl}}
}

