% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{RankControlHamming-class}
\alias{RankControlHamming}
\alias{RankControlHamming-class}
\title{RankControlHamming Class}
\description{
A S4 class for the Hamming distance model fitting.
It is derived from class \code{\link{RankControl-class}}.
}

