% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{RankControlSpearman-class}
\alias{RankControlSpearman}
\alias{RankControlSpearman-class}
\title{RankControlSpearman Class}
\description{
A S4 class for the Spearman distance model fitting.
It is derived from class \code{\link{RankControl-class}}.
}

