% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{RanktoHash}
\alias{RanktoHash}
\title{Create Hash Value for Ranking}
\usage{
RanktoHash(r)
}
\arguments{
\item{r}{a vector or matrix of rankings. Each row of the matrix represents a ranking.
The ranking should be a integer from one to number of objects. No NA is allowed}
}
\value{
a vector of character strings representing the hash values.
}
\description{
Sometimes it is handy to deal with rankings as a hash value. \code{RanktoHash} returns hash values for ranks.
Maximum 52 objects are supported.
}
\seealso{
\code{\link{HashtoRank}} for a reverse operation.
}

