% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{DistanceBlock}
\alias{DistanceBlock}
\title{Calculate Kendall distance between one ranking and a matrix of rankings}
\usage{
DistanceBlock(mat, r)
}
\arguments{
\item{mat}{a matrix of rankings. Each row stores a ranking.}

\item{r}{a single ranking}
}
\value{
a vector of Kendall distances
}
\description{
Calculate Kendall distance between one ranking and a matrix of rankings
}

