% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{GenerateExample}
\alias{GenerateExample}
\title{Generate simple examples}
\usage{
GenerateExample(ranking = TRUE, central = 1:5, lambda = 0.2)
}
\arguments{
\item{ranking}{TRUE if "ranking" representation is used in the output data; otherwise "ordering" representation is used.}

\item{central}{The modal ranking.}

\item{lambda}{The parameter in the model.}
}
\description{
This function generates simple examples for illustrative proposes.
The sample contains rankings of five objects and the underlying model is a Mallows' phi
model with default dispersion parameter set to 0.2
and modal ranking set to (1,2,3,4,5)
}

