% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rankdist.R
\name{GenerateExampleTopQ}
\alias{GenerateExampleTopQ}
\title{Generate simple examples of top-q rankings}
\usage{
GenerateExampleTopQ(central = c(1, 2, 3, 4, 4), w = c(0.7, 0.5, 0.3, 0))
}
\arguments{
\item{central}{The modal ranking.}

\item{w}{The weights in the model.}
}
\description{
This function generates simple examples for illustrative proposes.
The sample contains the top-3 rankings of five objects and the underlying model is a weighted Kendall distance model
with default weights set to (0.7,0.5,0.3,0)
and modal ranking set to (1,2,3,4,4)
}

