% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{RankControlWtau-class}
\alias{RankControlWdbm}
\alias{RankControlWdbm-class}
\alias{RankControlWtau-class}
\title{RankControlWdbm Class}
\description{
A S4 class for the Weighted tau model fitting.
It is derived from class \code{\link{RankControl-class}}.
}
\section{Slots}{

\describe{
\item{\code{optimx_control}}{a list to be passed to \code{\link[optimx]{optimx}}. The list must not contain a component \code{maximize=TRUE} since internally the negation of the likelihood function is minimized.}
}}
\seealso{
\code{\link{RankData}}, \code{\link{RankInit}}, \code{\link{RankControl}}
}

